Rewrite `addEdge` to use binary search to insert the element in the already sorted slice.

Benchmark for `Insert`:

```
name       old time/op    new time/op    delta
Insert-16     624ns ± 4%     523ns ± 3%  -16.29%  (p=0.029 n=4+4)

name       old alloc/op   new alloc/op   delta
Insert-16      169B ± 0%      137B ± 0%  -18.93%  (p=0.029 n=4+4)

name       old allocs/op  new allocs/op  delta
Insert-16      4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.029 n=4+4)
```

Closes #15