@bep 2022-06-29
@armon It would be good if you could either close or merge this sooner rather than later. To me this looks like an obvious improvement, but up to you.